Support Lossless IPsec Transport Mode Rekey Input
Change the IPsec Transport mode policy for input SAs to
not use the SPI. Since the packet is already decoded and
simply matches a single transport mode policy on input,
we don't have any issue identifying the policy for removal,
and we don't have a way to support multiple policies in
parallel. To support packets on a new SA before the old SA
is deleted, we can't require a match on the SPI.

So, just remove the SPI, only on input, where it's not used
to actually look up a specific policy.

Bug: 270841813
Test: atest XfrmControllerParameterizedTest#\
        TestIpSecApplyTransportModeTransform
Flag: EXEMPT bugfix
Change-Id: I58bfedd8c2ffa09c93366a71be2d2a4d31b965f1